+++ cms/extensions/command-palette-module/src/command-palette/registry/index.ts
@@ -1,0 +1,4 @@
+export { defineCommands, registerCommands } from './register-commands';
+
+export type { CommandConfig, GroupConfig } from './types';
+export { useRegisteredCommands } from './use-registered-commands';