+++ cms/extensions/super-header-interface/src/composables/useFlows.ts
@@ -1,0 +1,139 @@
+import type { FlowIdentifier } from '../types';
+import { useApi, useStores } from '@directus/extensions-sdk';
+import { inject, ref } from 'vue';
+import { useI18n } from 'vue-i18n';
+import { useRoute } from 'vue-router';
+
+const showForm = ref(false);
+const currentFlow = ref<any>(null);
+
+export function useFlows(collection: string, primaryKey: string | number | null) {
+	const { t } = useI18n();
+	const api = useApi();
+	const formValues = inject('values') as Record<string, any>;
+	const route = useRoute();
+
+	const { useNotificationsStore } = useStores();
+	const notificationStore = useNotificationsStore();
+
+	const flowsCache = ref<Record<string, any>>({});
+	const loading = ref(false);
+
+	const fetchFlows = async (flowIdentifiers: FlowIdentifier[]) => {
+		const uniqueFlowIds = flowIdentifiers.filter(({ key }) => !flowsCache.value[key]);
+		if (uniqueFlowIds.length === 0)
+			return;
+
+		loading.value = true;
+
+		try {
+			const response = await api.get('/flows', {
+				params: {
+					filter: {
+						id: {
+							_in: uniqueFlowIds.map((flow) => flow.key),
+						},
+					},
+				},
+			});
+
+			response.data.data.forEach((flow: any) => {
+				flowsCache.value[`${flow.id}`] = flow;
+			});
+		}
+		catch (error) {
+			console.error('Error fetching flows:', error);
+		}
+		finally {
+			loading.value = false;
+		}
+	};
+
+	const getFlow = async (flowIdentifier: FlowIdentifier) => {
+		const cacheKey = flowIdentifier.key;
+
+		if (!flowsCache.value[cacheKey]) {
+			await fetchFlows([flowIdentifier]);
+		}
+
+		return flowsCache.value[cacheKey];
+	};
+
+	const showFlowForm = async (flow: any) => {
+		currentFlow.value = flow;
+		showForm.value = true;
+
+		return new Promise((resolve, reject) => {
+			const onSubmit = (values: Record<string, any>) => {
+				showForm.value = false;
+				resolve(values);
+			};
+
+			const onCancel = () => {
+				showForm.value = false;
+				reject(new Error('Flow cancelled'));
+			};
+
+			currentFlow.value = { ...flow, onSubmit, onCancel };
+		});
+		return {};
+	};
+
+	const runFlow = async (flowIdentifier: FlowIdentifier, values: Record<string, any>) => {
+		try {
+			const flow = await getFlow(flowIdentifier);
+			let formData = values as any;
+
+			if (flow.options?.requireConfirmation) {
+				formData = await showFlowForm(flow);
+			}
+
+			const response = await api.post(`/flows/trigger/${flow.id}`, {
+				...formData,
+				collection,
+				keys: primaryKey ? [primaryKey] : [],
+				$values: formValues.value,
+				$route: {
+					params: route.params,
+					query: route.query,
+					fullPath: route.fullPath,
+				},
+			});
+
+			notificationStore.add({
+				title: t('run_flow_success', { flow: flow.name }),
+			});
+
+			return response.data;
+		}
+		catch (error) {
+			console.error(`Error running flow ${flowIdentifier.key}:`, error);
+
+			notificationStore.add({
+				title: t('unexpected_error'),
+				type: 'error',
+				code: 'UNKNOWN',
+				dialog: true,
+				error,
+			});
+		}
+	};
+
+	const submitFlow = async (values: Record<string, any>) => {
+		if (!currentFlow.value)
+			return;
+		currentFlow.value.onSubmit(values);
+	};
+
+	return {
+		flowsCache,
+		loading,
+		fetchFlows,
+		getFlow,
+		runFlow,
+		showForm,
+		currentFlow,
+		showFlowForm,
+		submitFlow,
+	};
+}